change folder name, cleanup code
