--- Filament tip testing/TipTest.gcode
@@ -1,87 +1,0 @@
-M83
-M109 S250
-G28
-G1 Z50
-T1
-G1 E15 F1000
-; material : PLA -> PLA
-;--------------------
-M220 B
-M220 S100
-; CP TOOLCHANGE UNLOAD
-;WIDTH:0.600000
-G1  X77.187 Y109.793  
-M900 K0
-G1 F4405
-G1  X58.833  E0.4415
-G1 F4433
-G1  X40.364  E0.4443
-G1 F4628
-G1  X38.687  E0.0403
-G1  Y109.193  F7200
-G1 F4628
-G1  X56.293  E0.4235
-G1 F4962
-G1  X76.970  E0.4973
-G1 F5158
-G1  X77.187  E0.0052
-G1  Y108.593  F7200
-G1 F5158
-G1  X55.914  E0.5117
-G1 F5213
-G1  X38.687  E0.4144
-G1  Y107.993  F7200
-G1 F5213
-G1  X43.182  E0.1081
-G1 F5381
-G1  X65.602  E0.5392
-G1 F5966
-G1  X77.187  E0.2787
-G1  Y107.393  F7200
-G1 F5966
-G1  X63.914  E0.3193
-G1 F7249
-G1  X38.687  E0.6068
-G1  Y106.793  F7200
-G1 F7249
-G1  X43.662  E0.1197
-G1 F8949
-G1  X77.187  E0.8064
-G1  Y106.193  F7200
-G1 F8949
-G1  X73.424  E0.0905
-G1 F10259
-G1  X38.687  E0.8355
-G1  Y105.593  F7200
-G1 F10259
-G1  X46.698  E0.1927
-G1 F10427
-G1  X77.187  E0.7334
-G1  Y104.993  F7200
-G1 F10427
-G1  X64.231  E0.3116
-G1 F9061
-G1  X38.687  E0.6144
-G1  Y104.393  F7200
-G1 F9061
-G1  X50.895  E0.2936
-G1 F7165
-G1  X77.187  E0.6324
-M73 P18 R19
-G1  Y103.793  F7200
-G1 F7165
-G1  X73.625  E0.0857
-;WIDTH:0.500000
-G1 E-15.0000 F6000
-G1 E-8.7500 F1200
-G1 E-2.5000 F600
-G1 E-1.2500 F360
-M104 S200
-G1  Y103.193 
-G1  X38.687  E5.0000 F1258
-G1  X73.625  E-5.0000 F839
-G1 E-2.5000 F2000
-G1  Y103.243  F2400
-G4 S0
-; Filament-specific end gcode 
-G1 E-100
