+++ FilamentTipTesting/TipTest.gcode
@@ -1,0 +1,115 @@
+;Script generated Gcode begin
+M83
+M109 S320
+G28
+G1 Z50
+T1
+G1 E15 F1000
+;Script generated Gcode end
+
+;Slicer generated Gcode begin
+; material : PLA -> PLA
+;--------------------
+M220 B
+M220 S100
+; CP TOOLCHANGE UNLOAD
+;WIDTH:0.6
+G1  X219.250 Y189.250  
+M900 K0
+G1 F3554
+G1  X204.441  E0.686
+G1 F3662
+G1  X189.182  E0.7068
+G1 F3877
+G1  X173.027  E0.7484
+G1 F4093
+G1  X160.750  E0.5687
+G1  Y188.650  F7200
+G1 F4093
+G1  X165.526  E0.2212
+G1 F4254
+G1  X183.253  E0.8211
+G1 F4416
+G1  X201.652  E0.8523
+G1 F4685
+G1  X219.250  E0.8151
+G1  Y188.050  F7200
+G1 F4685
+G1  X217.326  E0.0891
+G1 F5062
+G1  X196.234  E0.977
+G1 F5331
+G1  X174.020  E1.029
+G1 F5385
+G1  X160.750  E0.6147
+G1  Y187.450  F7200
+G1 F5385
+G1  X169.918  E0.4247
+;WIDTH:0.5
+G1 E-15 F6000
+G1 E-55.3 F5400
+G1 E-15.8 F2700
+G1 E-7.9 F1620
+M104 S205
+G1  Y186.850 
+G1  X219.250  E5 F1302
+G1  X169.918  E-5 F1404
+G1  X219.250  E5 F1505
+G1  X169.918  E-5 F1607
+G1  X219.250  E5 F1708
+G1  X169.918  E-5 F1810
+G1  X219.250  E5 F1911
+G1  X169.918  E-5 F2013
+G1 E2 F2000
+G1  Y186.900  F2400
+G4 S0
+; Filament-specific end gcode 
+;Slicer generated Gcode end
+
+;Script generated Gcode begin
+G1 E-100
+;Script generated Gcode end
+
+; Original Gcode file start line: 2060
+; Original Gcode file end line: 2115
+; No filament profile name found
+; filament_wipe_advanced_pigment = 0.5,0.5,0.5,0.5,0.5
+; filament_vendor = 123-3D
+; filament_use_skinnydip = 0,0,0,0,0
+; filament_use_fast_skinnydip = 0,0,0,0,0
+; filament_unloading_speed_start = 100,100,100,100,100
+; filament_unloading_speed = 90,90,90,90,90
+; filament_unload_time = 0,0,0,0,0
+; filament_type = PLA;PLA;PLA;PLA;PLA
+; filament_toolchange_temp = 200,200,200,200,200
+; filament_toolchange_part_fan_speed = 50,50,50,50,50
+; filament_toolchange_delay = 0,0,0,0,0
+; filament_spool_weight = 0,0,0,0,0
+; filament_soluble = 0,0,0,0,0
+; filament_skinnydip_distance = 31,31,31,31,31
+; filament_shrink = 100%,100%,100%,100%,100%
+; filament_settings_id = "123-3D Jupiter PLA @CREALITY";"123-3D Jupiter PLA @CREALITY";"123-3D Jupiter PLA @CREALITY";"123-3D Jupiter PLA @CREALITY";"123-3D Jupiter PLA @CREALITY"
+; filament_ramming_parameters = "120 100 6.6 6.8 7.2 7.6 7.9 8.2 8.7 9.4 9.9 10.0| 0.05 6.6 0.45 6.8 0.95 7.8 1.45 8.3 1.95 9.7 2.45 10 2.95 7.6 3.45 7.6 3.95 7.6 4.45 7.6 4.95 7.6";"120 100 6.6 6.8 7.2 7.6 7.9 8.2 8.7 9.4 9.9 10.0| 0.05 6.6 0.45 6.8 0.95 7.8 1.45 8.3 1.95 9.7 2.45 10 2.95 7.6 3.45 7.6 3.95 7.6 4.45 7.6 4.95 7.6";"120 100 6.6 6.8 7.2 7.6 7.9 8.2 8.7 9.4 9.9 10.0| 0.05 6.6 0.45 6.8 0.95 7.8 1.45 8.3 1.95 9.7 2.45 10 2.95 7.6 3.45 7.6 3.95 7.6 4.45 7.6 4.95 7.6";"120 100 6.6 6.8 7.2 7.6 7.9 8.2 8.7 9.4 9.9 10.0| 0.05 6.6 0.45 6.8 0.95 7.8 1.45 8.3 1.95 9.7 2.45 10 2.95 7.6 3.45 7.6 3.95 7.6 4.45 7.6 4.95 7.6";"120 100 6.6 6.8 7.2 7.6 7.9 8.2 8.7 9.4 9.9 10.0| 0.05 6.6 0.45 6.8 0.95 7.8 1.45 8.3 1.95 9.7 2.45 10 2.95 7.6 3.45 7.6 3.95 7.6 4.45 7.6 4.95 7.6"
+; filament_notes = ;;;;
+; filament_minimal_purge_on_wipe_tower = 15,15,15,15,15
+; filament_melt_zone_pause = 0,0,0,0,0
+; filament_max_wipe_tower_speed = 0,0,0,0,0
+; filament_max_volumetric_speed = 15,15,15,15,15
+; filament_max_speed = 0,0,0,0,0
+; filament_max_overlap = 100%,100%,100%,100%,100%
+; filament_loading_speed_start = 3,3,3,3,3
+; filament_loading_speed = 28,28,28,28,28
+; filament_load_time = 0,0,0,0,0
+; filament_enable_toolchange_temp = 0,0,0,0,0
+; filament_enable_toolchange_part_fan = 0,0,0,0,0
+; filament_dip_insertion_speed = 33,33,33,33,33
+; filament_dip_extraction_speed = 70,70,70,70,70
+; filament_diameter = 1.75,1.75,1.75,1.75,1.75
+; filament_density = 1.24,1.24,1.24,1.24,1.24
+; filament_custom_variables = ;;;;
+; filament_cost = 19.5,19.5,19.5,19.5,19.5
+; filament_cooling_zone_pause = 0,0,0,0,0
+; filament_cooling_moves = 4,4,4,4,4
+; filament_cooling_initial_speed = 2.2,2.2,2.2,2.2,2.2
+; filament_cooling_final_speed = 3.4,3.4,3.4,3.4,3.4
+; filament_colour = #FFE200;#FFE200;#FFE200;#FFE200;#FFE200
