+++ Settings/Creality Style Hotend/Creality Style Hotend PLA.ini
@@ -1,0 +1,77 @@
+[filament:Creality Style Hotend PLA]
+bed_temperature = 60
+bridge_fan_speed = 30
+chamber_temperature = 0
+compatible_printers = 
+compatible_printers_condition = 
+compatible_prints = 
+compatible_prints_condition = 
+cooling = 1
+disable_fan_first_layers = 3
+end_filament_gcode = "; Filament-specific end gcode \n;END gcode for filament\n"
+external_perimeter_fan_speed = -1
+extrusion_multiplier = 1
+fan_always_on = 1
+fan_below_layer_time = 60
+filament_colour = #29B2B2
+filament_cooling_final_speed = 10
+filament_cooling_initial_speed = 10
+filament_cooling_moves = 1
+filament_cooling_zone_pause = 0
+filament_cost = 0
+filament_density = 0
+filament_deretract_speed = nil
+filament_diameter = 1.75
+filament_dip_extraction_speed = 90
+filament_dip_insertion_speed = 90
+filament_enable_toolchange_part_fan = 1
+filament_enable_toolchange_temp = 1
+filament_load_time = 12
+filament_loading_speed = 14
+filament_loading_speed_start = 19
+filament_max_speed = 0
+filament_max_volumetric_speed = 0
+filament_max_wipe_tower_speed = 0
+filament_melt_zone_pause = 0
+filament_minimal_purge_on_wipe_tower = 15
+filament_notes = ""
+filament_ramming_parameters = "120 100 6.53226 6.64516 6.70968 6.74194 6.85484 6.91935 6.85484 7.51613 10.0645 13.5 15.3387 14.9516 13.5161 12.0645 11.1129 10.7581| 0.05 6.49249 0.45 6.69249 0.95 6.74623 1.45 6.93442 1.95 7.97949 2.45 14.957 2.95 14.3848 3.45 11.4387 3.95 10.686 4.45 9.6 4.95 7.37419"
+filament_retract_before_travel = nil
+filament_retract_before_wipe = nil
+filament_retract_layer_change = nil
+filament_retract_length = nil
+filament_retract_lift = nil
+filament_retract_lift_above = nil
+filament_retract_lift_below = nil
+filament_retract_restart_extra = nil
+filament_retract_speed = nil
+filament_settings_id = "Creality Style Hotend PLA"
+filament_shrink = 100%
+filament_skinnydip_distance = 47
+filament_soluble = 0
+filament_spool_weight = 0
+filament_toolchange_delay = 0
+filament_toolchange_part_fan_speed = 50
+filament_toolchange_temp = 160
+filament_type = PLA
+filament_unload_time = 12
+filament_unloading_speed = 80
+filament_unloading_speed_start = 120
+filament_use_fast_skinnydip = 0
+filament_use_skinnydip = 1
+filament_vendor = (Unknown)
+filament_wipe = nil
+filament_wipe_advanced_pigment = 0.5
+filament_wipe_extra_perimeter = nil
+first_layer_bed_temperature = 60
+first_layer_temperature = 180
+full_fan_speed_layer = 0
+inherits = 
+max_fan_speed = 30
+max_speed_reduction = 90%
+min_fan_speed = 30
+min_print_speed = 10
+slowdown_below_layer_time = 5
+start_filament_gcode = "; Filament gcode\n"
+temperature = 180
+top_fan_speed = -1
